implemented dynamic uniforms (for using different uniform values across different draw calls)

+++ include/fae/rendering/mesh.hpp
@@ -18,7 +18,7 @@ namespace fae
     struct mesh
     {
         std::vector<vertex> vertices;
-        std::vector<std::size_t> indices;
+        std::vector<std::uint32_t> indices;
 
         static auto load(std::filesystem::path path) -> std::optional<mesh>;
